feat: Logbat에서 처리 불가능한 예외는 DEBUG로 설정하여 전송하도록 리팩토링함
