+++ sdk/java/logbat-sdk/src/main/java/info/logbat/application/LogSendRequestFactory.java
@@ -1,5 +1,6 @@
 package info.logbat.application;
 
+import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.spi.ILoggingEvent;
 import info.logbat.infrastructure.payload.LogSendRequest;
 import java.time.Instant;
@@ -7,17 +8,27 @@ import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
 import java.util.Map;
+import java.util.Set;
 
 public class LogSendRequestFactory {
 
     private static final DateTimeFormatter CUSTOM_FORMATTER =
         DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss");
+    /**
+     * Logbat에서 기본적으로 지원하는 로그 레벨
+     */
+    private static final Set<Level> SUPPORTED_LEVELS = Set.of(
+        Level.TRACE, Level.DEBUG, Level.INFO, Level.WARN, Level.ERROR
+    );
+    private static final String OTHER_LEVEL = "DEBUG";
 
     /**
      * ILoggingEvent로부터 LogSendRequest를 생성하는 메서드
      */
     public LogSendRequest createLogSendRequest(ILoggingEvent eventObject) {
-        String formattedLevel = eventObject.getLevel().toString();
+
+        Level level = eventObject.getLevel();
+        String formattedLevel = getFormattedLevel(level);
         String timestamp = LocalDateTime.ofInstant(
             Instant.ofEpochMilli(eventObject.getTimeStamp()),
             ZoneId.systemDefault()
@@ -41,4 +52,13 @@ public class LogSendRequestFactory {
 
         return new LogSendRequest(formattedLevel, message, timestamp);
     }
+
+    private String getFormattedLevel(Level level) {
+        return isSupportedLevel(level) ? level.toString() : OTHER_LEVEL;
+    }
+
+    private boolean isSupportedLevel(Level level) {
+        return SUPPORTED_LEVELS.contains(level);
+    }
+
 }
